ci: add ruff/ty CI workflow and fix lint/type issues
- Create .github/workflows/ci.yml with ruff check, ruff format, ty check, pytest
- Add ruff to dev dependencies in pyproject.toml
- Remove stale type: ignore comments in server/__init__.py
- Fix is_configured() -> _is_configured() in tests (method is private)
- Fix is_integer type narrowing in datetime_utils.py
- Suppress false positive dict.get arg-type in tool_config.py

+++ src/toolregistry_hub/datetime_utils.py
@@ -160,7 +160,7 @@ class DateTime:
         # Calculate time difference
         source_offset = source_time.utcoffset() or timedelta()
         target_offset = target_time.utcoffset() or timedelta()
-        hours_difference = (target_offset - source_offset).total_seconds() / 3600
+        hours_difference = float((target_offset - source_offset).total_seconds() / 3600)
 
         # Format time difference string
         if hours_difference.is_integer():
